Holoscan SDK v3.11.0 Release

+++ examples/gpu_resident_example/gpu_resident_example.cpp
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+ * SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -326,7 +326,7 @@ int main() {
   }
 
   // Run 10 iterations to test the GPU resident execution
-  unsigned int seed = time(nullptr);
+  unsigned int seed = static_cast<unsigned int>(time(nullptr));
   for (int iteration = 0; iteration < 10; ++iteration) {
     if (!run_iteration(source_op, sink_op, gr_fragment, seed, iteration)) {
       break;
